Add regression and feature tasks
