+++ build.gradle
@@ -46,6 +46,16 @@ task checkTests(type: GradleBuild) {
     tasks = ['testDebugUnitTest', 'connectedDebugAndroidTest']
 }
 
+task checkFeatures(type: GradleBuild) {
+    tasks = ["connectedDebugAndroidTest"]
+    startParameter.projectProperties = ["android.testInstrumentationRunnerArguments.notPackage": "org.odk.collect.android.regression"]
+}
+
+task checkRegression(type: GradleBuild) {
+    tasks = ["connectedDebugAndroidTest"]
+    startParameter.projectProperties = ["android.testInstrumentationRunnerArguments.package": "org.odk.collect.android.regression"]
+}
+
 ext {
     daggerVersion = '2.27'
     okhttp3Version = '4.4.1'
